+++ tests/Support/Stubs/Controllers/ExceptionController.php
@@ -3,10 +3,16 @@
 namespace Somnambulist\ApiBundle\Tests\Support\Stubs\Controllers;
 
 use Assert\InvalidArgumentException;
+use Exception;
+use http\Env;
+use RuntimeException;
 use Somnambulist\ApiBundle\Controllers\ApiController;
 use Somnambulist\ApiBundle\Tests\Support\Stubs\Entities\MyAssertingEntity;
 use Somnambulist\Domain\Entities\Exceptions\EntityNotFoundException;
 use Somnambulist\Domain\Entities\Exceptions\InvalidDomainStateException;
+use stdClass;
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Exception\HandlerFailedException;
 
 /**
  * Class ExceptionController
@@ -47,5 +53,17 @@ class ExceptionController extends ApiController
     public function assertAction()
     {
         new MyAssertingEntity('', '', '', '');
+    }
+
+    /**
+     * @throws HandlerFailedException
+     */
+    public function messengerAction()
+    {
+        try {
+            new MyAssertingEntity('', '', '', '');
+        } catch (Exception $e) {
+            throw new HandlerFailedException(new Envelope(new stdClass()), [$e]);
+        }
     }
 }